[UNTESTED] Fix integer overflow when reading late object properties in big files. Better V11 CustomItemDatas redirection support. (#6)
* Fix an int overflow when loading object properties beyond 2GB mark
* Provide SaveVersion through ArkArchive to rest of the code
* Read CustomItemDatas as native structs when version is higher than 11
Object and classes references are suspiciously missing from ArkCryoStore, which was crucial to figuring this out - thanks @miragedmuk.

+++ SavegameToolkit/GameObject.cs
@@ -280,7 +280,7 @@ namespace SavegameToolkit {
         }
 
         public void LoadProperties(ArkArchive archive, GameObject next, int propertiesBlockOffset) {
-            int offset = propertiesBlockOffset + propertiesOffset;
+            long offset = propertiesBlockOffset + propertiesOffset;
             long nextOffset = propertiesBlockOffset + next?.propertiesOffset ?? archive.Limit;
 
             archive.Position = offset;
